Move multiInputWrapper.js to wrapper directory
